refactor: extract funding_source
