+++ lnbits/core/services/services.py
@@ -8,15 +8,8 @@ from lnbits.settings import (
     readonly_variables,
     settings,
 )
-from lnbits.wallets import get_funding_source, set_funding_source
 
-from ..crud import (
-    get_total_balance,
-    update_admin_settings,
-)
-from ..models import (
-    BalanceDelta,
-)
+from ..crud import update_admin_settings
 
 
 async def check_webpush_settings():
@@ -54,21 +47,3 @@ def update_cached_settings(sets_dict: dict):
             logger.warning(f"Failed overriding setting: {key}, value: {value}")
     if "super_user" in sets_dict:
         settings.super_user = sets_dict["super_user"]
-
-
-async def switch_to_voidwallet() -> None:
-    funding_source = get_funding_source()
-    if funding_source.__class__.__name__ == "VoidWallet":
-        return
-    set_funding_source("VoidWallet")
-    settings.lnbits_backend_wallet_class = "VoidWallet"
-
-
-async def get_balance_delta() -> BalanceDelta:
-    funding_source = get_funding_source()
-    status = await funding_source.status()
-    lnbits_balance = await get_total_balance()
-    return BalanceDelta(
-        lnbits_balance_msats=lnbits_balance,
-        node_balance_msats=status.balance_msat,
-    )
